+++ src/Types/Audio.php
@@ -45,9 +45,9 @@ class Audio extends BaseType
     public ?string $mime_type = null;
     /**
      * Optional. File size
-     * @var int|null
+     * @var float|null
      */
-    public ?int $file_size = null;
+    public ?float $file_size = null;
     /**
      * Optional. Thumbnail of the album cover to which the music file belongs
      * @var PhotoSize|null
