+++ src/Types/Document.php
@@ -40,9 +40,9 @@ class Document extends BaseType
     public ?string $mime_type = null;
     /**
      * Optional. File size
-     * @var int|null
+     * @var float|null
      */
-    public ?int $file_size = null;
+    public ?float $file_size = null;
 
     protected function bindObjects($key, $data): ?object
     {
